Viðbót í viku 2

+++ Vika2/FTilC.java
@@ -5,7 +5,7 @@ import java.util.Scanner;
  *  Nafn    : Ebba Þóra Hvannberg
  *  T-póstur: ebba@hi.is
  *
- *  Lýsing  : Tekur inn hitastig í Fahrenheit (heiltala) frá console
+ *  Lýsing  : Tekur inn hitastig í Fahrenheit (kommutala) frá console
  *  og skilar út í Celcíus á staðalúttaki
  *
  *****************************************************************************/
@@ -13,8 +13,8 @@ import java.util.Scanner;
 public class FTilC {
     public static void main(String[] args) {
         Scanner inntak = new Scanner(System.in, StandardCharsets.UTF_8);
-        int fahrenheit = inntak.nextInt();
-        int celcius = (fahrenheit - 32) * 5 / 9;
+        double fahrenheit = inntak.nextDouble();
+        double celcius = (fahrenheit - 32) * 5 / 9;
         System.out.println(celcius);
     }
 }
